Create delete_post.php
I have the delete part set up. Only the author of a post can delete it. But even though the "delete" button does not work for the other post, it is still displayed. We have to fix on hiding it for all but the post written by the user online.