Update box.php
The delete/edit button are now only displayed if the $_SESSION["user_id"] equals the user_id saved in the posts table! That means if you did not make the post, the buttons are not displayed and on the other hand, you are able to keep track of which posts are yours and you get the chance to delete them.