Update signup.php
Since we are using 'pattern' to validate the input into our form fields, I took out the other form of validation. The password/confirm password had to stay in though because it still has to be checked if they are the same. I also changed the echoed message to an alert message.